requirements.txt
fastapi
uvicorn
gunicorn
jinja2
email-validator
python-multipart
passlib
python-jose
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
authlib
httpx
itsdangerous